skills/ketchup-diff-analyzer/requirements.txt
# Ketchup Diff Analyzer Dependencies
GitPython>=3.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
